Refactor of components requested on pull request
