+++ src/hooks/useMountTransition.tsx
@@ -1,0 +1,23 @@
+import React, { useState, useEffect } from "react"
+
+const useMountTransition = (isMounted: boolean, unmountDelay: number) => {
+  const [hasTransitionedIn, setHasTransitionedIn] = useState(false)
+
+  useEffect(() => {
+    let timeoutId: string | number | NodeJS.Timeout | undefined
+
+    if (isMounted && !hasTransitionedIn) {
+      setHasTransitionedIn(true)
+    } else if (!isMounted && hasTransitionedIn) {
+      timeoutId = setTimeout(() => setHasTransitionedIn(false), unmountDelay)
+    }
+
+    return () => {
+      if (timeoutId) clearTimeout(timeoutId)
+    }
+  }, [unmountDelay, isMounted, hasTransitionedIn])
+
+  return hasTransitionedIn
+}
+
+export default useMountTransition
