+++ src/styles/photoDetail.scss
@@ -223,8 +223,6 @@
       margin: 0 auto;
       width: 70%;
       max-height: 60%;
-      // transform: scale(1);
-      //transition: transform 1s ease-in-out;
       z-index: 20;
       cursor: zoom-in;
     }
